Controle de Usuario

+++ app/Controllers/Http/CashMovementController.js
@@ -5,16 +5,17 @@ const Cashier = use('App/Models/Cashier')
 
 class CashMovementController {
 
-  async index () {
-    const cashMovement = CashMovement.all()
+  async index ({auth}) {
+    const { id } = auth.user
+    const cashMovement = CashMovement.query().where({user_id:id}).fetch();
     return cashMovement
   }
 
-  async store ({ auth, request }) {
+  async store ({ auth, request, response }) {
     const { id } = auth.user
     const data = request.only([ 'cashier_id','categories_id','type','value','description'])
     if (!await Database.table('cashiers').where({id:data.cashier_id}).where({user_id: id}).first()){
-      return 'Erro, Caixa não liberado para este Usuario'
+      return response.status(401).send({ error: 'Not authorized' })
     }
     /**Grava Registro no BD */
     const cashMovement = await CashMovement.create({...data, user_id: id})
@@ -24,6 +25,7 @@ class CashMovementController {
     const type = request.input('type')
     const  cashier = await Database.select('saldo').from('cashiers').where({id: cashier_id})
     var saldo = cashier[0]['saldo']
+
     if ((type == 'E') || (type == 'e')){
         saldo = saldo + value
     } if ((type == 'S') || (type == 's')) {
@@ -35,15 +37,21 @@ class CashMovementController {
     return cashMovement
   }
   
-  async show ({ params }) {
+  async show ({ params, auth }) {
     const cashMovement = await CashMovement.findOrFail(params.id)
+    if (cashMovement.user_id !== auth.user.id){
+      return response.status(401).send({ error: 'Not authorized' })
+    }
     return cashMovement
   }
 
-  async destroy ({ params }) {
+  async destroy ({ params, auth, response}) {
     const cashMovement = await CashMovement.findOrFail(params.id)
+    if (cashMovement.user_id !== auth.user.id){
+      return response.status(401).send({ error: 'Not authorized' })
+    }
     await cashMovement.delete()
-    return cashMovement
+    return cashMovement.id
   }
 }
